JIT: Extract StructSegments to its own file
Also move the function to compute the significant segments for a class
handle from `Promotion` to `Compiler`.
