+++ src/coreclr/jit/CMakeLists.txt
@@ -178,6 +178,7 @@ set( JIT_SOURCES
   ssabuilder.cpp
   ssarenamestate.cpp
   stacklevelsetter.cpp
+  structsegments.cpp
   switchrecognition.cpp
   treelifeupdater.cpp
   unwind.cpp
@@ -379,6 +380,7 @@ set( JIT_HEADERS
   ssaconfig.h
   ssarenamestate.h
   stacklevelsetter.h
+  structsegments.h
   target.h
   targetx86.h
   targetamd64.h
